refactor: Update file_writer_spec.rb and writer_manager.rb
Refactor the file_writer_spec.rb and writer_manager.rb files to simplify the file writing and updating logic.
